Upgrade to 2022.1, rework VDMA synchronization
The previous VDMA synchronization setup had many issues, among which a
lot of frame tearing and the requirement that pcam A be the
synchronization master. We reconfigured all of it to support any camera
being master as well as to prevent frame tearing, as long as all
connected pcams run normally (i.e. none of them gets abnormally stuck on
a frame or, conversely, no pcam is ever more than one frame behing the
master). In short, all VDMA write channels are configured as slaves in
the hardware, of which one is selected at runtime to be promoted to a
"master", by disabling its genlock functionality, making it cycle
through the framebuffers like a blind master. All other slaves follow
it. More details on this structure can be found in the project's
documentation page:
https://digilent.com/reference/programmable-logic/zedboard/demos/fmc-pcam