Fix project_info to contain updated board_part
The 2022.1 upgrade required changing the project board_part from
`em.avnet.com:zed:part0:1.4` to `avnet.com:zedboard:part0:1.4`. At the
time of building and exporting the hardware platform, this change was
applied to the project's state, but the checkin script failed to also
save it because of the DOCUMENTED behaviour that project_info.tcl is not
recreated if it already exists. This commit adds this change to
project_info.tcl but WILL NOT be followed by another hardware export
commit.